Added Familiar, Farming Village, Hunting Party, Ironworks, Moneylender, Sea Hag, and Treasure Map to basicAI's actionPriority, with following positions:
Farming Village - highest priority drawing Village
Familiar - last of nonterminal card drawers, in case a King's Court or Throne Room can be drawn before it is played
Hunting Party - right above Lab and Alchemist; a tricky one to place just right, so I went with cards it's similar to
Treasure Map - if it will gain Golds, it's the highest priority terminal. Otherwise, it's priority is almost at the bottom, and will only play one Map if its the only one in the deck and the deck already has 4 Golds.
Sea Hag - after Mountebank and Witch so the benefits those cards give will be taken advantage of first
Moneylender - after Steward to favor faster trashing
Ironworks - a hard one to place; for now I put it just above Workshop. If it's going to be terminal, I think that's an acceptable spot, but it should be higher if it's going to gain an Action card