Treasure Map was broken. The line I originally wrote for trashing the second map was:
doTrash(c['Treasure Map'])

which was changed to:

state.doTrash(state.current, c['Treasure Map']),

but this was still incorrect. It is now:

state.current.doTrash(c['Treasure Map'])

I discovered it was broken by testing each card I've submitted on the live site, and when the AI got to a hand with two T Maps, it played one then the simulator just stopped working.

I have also added code so it will properly deal with the case where the number of Golds in the supply is less than 4, and it will output to the log the trashing of both Maps.